teppi runtime on del

+++ code/modules/mob/living/simple_mob/subtypes/animal/alien animals/teppi.dm
@@ -687,7 +687,9 @@ GLOBAL_VAR_INIT(teppi_count, 0)	// How mant teppi DO we have?
 			var/nutrition_cost = 500 + (nutrition / 2)
 			adjust_nutrition(-nutrition_cost)
 			new /mob/living/simple_mob/vore/alienanimals/teppi(loc, src)
-			qdel(src)
+			if(!QDELETED(src))
+				qdel(src)
+				return
 		else
 			visible_message("\The [src] whines pathetically...", runemessage = "whines")
 			if(prob(50))
